+++ lib/main.dart
@@ -139,7 +139,7 @@ Future<void> _handleStart() async {
 }
 
 String _buildNotificationTitle(Profile? profile) {
-  if (profile == null) return 'FlClashX';
+  if (profile == null) return 'Delore';
   final profileName = profile.label ?? profile.id;
 
   String serviceName = '';